feat: add shared agent tools module
